Файлы разбиты по классам. Адаптер связан с графикой.
