+++ Artifact.cpp
@@ -1,4 +1,4 @@
-#include "Things.h"
+#include "Artifact.h"
 
 Artifact::Artifact(std::string name, std::string id, short level, short power) : Item(name, id, level), power(power) {}
