+++ Located.cpp
@@ -1,4 +1,4 @@
-#include "Location.h"
+#include "Located.h"
 
 coord::coord(int x, int y) : x(x), y(y) {}
