+++ RGI/RGI.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 3.5.1, 2018-05-14T21:59:44. -->
+<!-- Written by QtCreator 3.5.1, 2018-05-16T01:00:06. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
