+++ libAdapter.h
@@ -1,14 +1,19 @@
-#ifndef RASSEKI_GAME_LIBADAPTER_H
-#define RASSEKI_GAME_LIBADAPTER_H
+#ifndef LIBADAPTER_H
+#define LIBADAPTER_H
 
 #include <vector>
 #include <string>
+
+#include "Storage.h"
+#include "Hero.h"
+
+Storage Inventory(5);
+Hero hero("MLQ", "Moleque", 5, 4, {1,1}, &Inventory);
 
 int Game();
 bool Save(std::string);
 void Load(std::string);
 std::vector<std::string> GetSavedNames();
+bool Go(int x, int y);
 
-int Go(int x, int y);
-
-#endif //RASSEKI_GAME_LIBADAPTER_H
+#endif //LIBADAPTER_H
